Module mini_haskell::utils::control

Control flow utilities.

Enums

Result3
Result type for parsing.

Void
The uninhabited type Void.

Traits

Either
The std::ops::Try trait is not yet stable. We roll up our own for now. It is named after Either, Left, and Right from Haskell.

Maybe
Named after Maybe, Just, and Nothing from Haskell. Use this for success/failure semantics, since Either is used to model the control flow.